Run replace_so_abi_tags on out of tree builds
As requested by henryiii.
